+++ src/main/java/menu/domain/coach/CoachRepository.java
@@ -1,8 +1,10 @@
-package menu.domain;
+package menu.domain.coach;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+
+import menu.domain.coach.Coach;
 
 public class CoachRepository {
