+++ src/main/java/menu/domain/coach/CoachService.java
@@ -1,14 +1,18 @@
-package menu.domain;
+package menu.domain.coach;
 
 import java.util.List;
 import java.util.Set;
+
+import menu.domain.coach.Coach;
+import menu.domain.coach.CoachName;
+import menu.domain.coach.CoachRepository;
 
 public class CoachService {
 
     public void saveCoach(List<String> coaches) {
         coaches.forEach(coach ->
                 CoachRepository.save(new Coach(
-                        new Name(coach)
+                        new CoachName(coach)
                         )
                 )
         );
