+++ src/main/java/menu/domain/menu/Category.java
@@ -1,4 +1,4 @@
-package menu.domain;
+package menu.domain.menu;
 
 import java.util.Arrays;
 import java.util.Optional;
