+++ src/main/java/menu/domain/menu/MenuRepository.java
@@ -1,6 +1,6 @@
-package menu.domain;
+package menu.domain.menu;
 
-import static menu.domain.Category.*;
+import static menu.domain.menu.Category.*;
 
 import java.util.ArrayList;
 import java.util.HashMap;
