+++ src/test/java/menu/domain/CoachTest.java
@@ -8,6 +8,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import menu.MenuConfig;
+import menu.domain.coach.Coach;
+import menu.domain.coach.CoachRepository;
+import menu.domain.menu.MenuRepository;
 
 class CoachTest {
